detector/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
psutil==5.9.8
pyyaml==6.0.1
